requirements.txt
RPi.GPIO==0.6.3
psycopg2==2.7.6.1